Rename LinkListener (#272)

+++ linklistener/linklistener.py
@@ -1,20 +1,19 @@
 import logging
 import re
 from io import BytesIO
-from typing import Optional
 
 import discord
 from contextlib2 import suppress
-from discordmenu.embed.components import EmbedAuthor, EmbedField, EmbedFooter, EmbedMain, EmbedThumbnail
+from discordmenu.embed.components import EmbedAuthor, EmbedFooter, EmbedMain
 from discordmenu.embed.view import EmbedView
 from redbot.core import commands
 
-logger = logging.getLogger('red.misc-cogs.replylistener')
+logger = logging.getLogger('red.misc-cogs.linklistener')
 
 LINK_REGEX = r'https?://discord.com/channels/(?:@me|\d+)/(\d+)/(\d+)'
 
 
-class ReplyListener(commands.Cog):
+class LinkListener(commands.Cog):
     """Turn message links into nice embeds"""
 
     def __init__(self, bot, *args, **kwargs):
@@ -38,12 +37,16 @@ class ReplyListener(commands.Cog):
         if await self.bot.cog_disabled_in_guild(self, message.guild):
             return
 
-        if (not (match := re.fullmatch(LINK_REGEX, message.content))) \
-                or message.attachments or message.embeds:
-            # Message is not just a link
+        if re.fullmatch(LINK_REGEX, message.content) \
+                and not message.attachments and not message.embeds:
+            just_link = True
+        elif re.search(LINK_REGEX, message.content):
+            just_link = False
+        else:
+            # message does not contain a link
             return
 
-        cid, mid = map(int, match.groups())
+        cid, mid = map(int, re.search(LINK_REGEX, message.content).groups())
 
         channel = self.bot.get_channel(cid)
         if channel is None:
@@ -56,8 +59,9 @@ class ReplyListener(commands.Cog):
         if not ref_message.content:
             return
 
-        with suppress(discord.Forbidden):
-            await message.delete()
+        if just_link:
+            with suppress(discord.Forbidden):
+                await message.delete()
 
         await message.channel.send(embed=self.message_to_embed(ref_message, message.author),
                                    reference=ref_message if ref_message.channel == message.channel else None,
